Fix `dllist_remove` when removing middle element
Within `dllist_remove` when we were removing middle element, there was no check for this.
So whenever user removed middle element, list got updated but the middle element which is used for faster
search remained the same which led to exceptions like:

  * `TypeError: Argument must be a dllistnode`
  * `ValueError: dllistnode does not belong to a list`

Or even to Segmentation Fault.

This fix moves check for change after the element removal and adding comparasion
of deleted element with current middle element to trigger middle update if needed.